pyproject.toml
[project]
name = "backendd"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "giovana-rosa", email = "giovana.rosaifc@gmail.com"},
]
dependencies = [
    "django>=5.2.3",
    "django-phonenumber-field[phonenumbers]>=8.1.0",
    "djangorestframework>=3.16.0",
]
requires-python = "==3.10.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
